+++ utilities/options/options_util_test.cc
@@ -761,6 +761,7 @@ TEST_F(OptionsUtilTest, WalDirInOptins) {
 }  // namespace ROCKSDB_NAMESPACE
 
 int main(int argc, char** argv) {
+  ROCKSDB_NAMESPACE::port::InstallStackTraceHandler();
   ::testing::InitGoogleTest(&argc, argv);
 #ifdef GFLAGS
   ParseCommandLineFlags(&argc, &argv, true);