+++ utilities/ttl/ttl_test.cc
@@ -895,6 +895,7 @@ TEST_F(TtlOptionsTest, LoadTtlMergeOperator) {
 
 // A black-box test for the ttl wrapper around rocksdb
 int main(int argc, char** argv) {
+  ROCKSDB_NAMESPACE::port::InstallStackTraceHandler();
   ::testing::InitGoogleTest(&argc, argv);
   return RUN_ALL_TESTS();
 }